Totem Magic! -3 effect totems --Regeneration: Redstone Block --Speed: Ice --Haste: Diamond Block
--Totem Topper: Totem Heart

Each totem is crafted using a diamond, a cerulae / crimtae wooden totem template (depending on the effect), and its specified ingredient, tossed into essence.

Totems can be stacked on top of each other, and when a totem topper is added to the top, the three totems below the topper will give their effects in a radius of 10 blocks. The strength of the effect depends on their position in the totem.

-New Biome: Sacred Crove
--Grows with crimtae and cerulae trees (yes, very creative, I know)
--The essence fountain (the only place to get totem essence) spawns here.
--Crimtae and cerulae wood can be built with, but also turn into the new base totems